Project5_HR_Back_End/requirements.txt
astroid==4.0.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
coverage==7.11.1
dill==0.4.0
dnspython==2.8.0
feedparser==6.0.12
Flask==3.1.2
flask-cors==6.0.1
html2text==2025.4.15
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
platformdirs==4.5.0
PyJWT==2.10.1
pylint==4.0.1
pymongo==4.15.4
realpython-reader==1.1.4
sgmllib3k==1.0.0
tomli==2.3.0
tomlkit==0.13.3
typing_extensions==4.15.0
Werkzeug==3.1.3
